@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -27,6 +27,10 @@ func (_m *InfoServiceClient) Projects(ctx context.Context, in *v1.ProjectInfoReq
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Projects")
+	}
 
 	var r0 *v1.ProjectInfoResponse
 	var r1 error
@@ -61,6 +65,10 @@ func (_m *InfoServiceClient) Tenants(ctx context.Context, in *v1.TenantInfoReque
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Tenants")
+	}
+
 	var r0 *v1.TenantInfoResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantInfoRequest, ...grpc.CallOption) (*v1.TenantInfoResponse, error)); ok {
@@ -83,13 +91,12 @@ func (_m *InfoServiceClient) Tenants(ctx context.Context, in *v1.TenantInfoReque
 	return r0, r1
 }
 
-type mockConstructorTestingTNewInfoServiceClient interface {
+// NewInfoServiceClient creates a new instance of InfoServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewInfoServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewInfoServiceClient creates a new instance of InfoServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewInfoServiceClient(t mockConstructorTestingTNewInfoServiceClient) *InfoServiceClient {
+}) *InfoServiceClient {
 	mock := &InfoServiceClient{}
 	mock.Mock.Test(t)
 
