@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type InfoServiceServer struct {
 // Projects provides a mock function with given fields: _a0, _a1
 func (_m *InfoServiceServer) Projects(_a0 context.Context, _a1 *v1.ProjectInfoRequest) (*v1.ProjectInfoResponse, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Projects")
+	}
 
 	var r0 *v1.ProjectInfoResponse
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *InfoServiceServer) Projects(_a0 context.Context, _a1 *v1.ProjectInfoRe
 func (_m *InfoServiceServer) Tenants(_a0 context.Context, _a1 *v1.TenantInfoRequest) (*v1.TenantInfoResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Tenants")
+	}
+
 	var r0 *v1.TenantInfoResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.TenantInfoRequest) (*v1.TenantInfoResponse, error)); ok {
@@ -66,13 +74,12 @@ func (_m *InfoServiceServer) Tenants(_a0 context.Context, _a1 *v1.TenantInfoRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewInfoServiceServer interface {
+// NewInfoServiceServer creates a new instance of InfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewInfoServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewInfoServiceServer creates a new instance of InfoServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewInfoServiceServer(t mockConstructorTestingTNewInfoServiceServer) *InfoServiceServer {
+}) *InfoServiceServer {
 	mock := &InfoServiceServer{}
 	mock.Mock.Test(t)
 
