@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -27,6 +27,10 @@ func (_m *NetworkTrafficServiceClient) Modified(ctx context.Context, in *v1.Netw
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -61,6 +65,10 @@ func (_m *NetworkTrafficServiceClient) Usage(ctx context.Context, in *v1.Network
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.NetworkUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.NetworkUsageRequest, ...grpc.CallOption) (*v1.NetworkUsageResponse, error)); ok {
@@ -83,13 +91,12 @@ func (_m *NetworkTrafficServiceClient) Usage(ctx context.Context, in *v1.Network
 	return r0, r1
 }
 
-type mockConstructorTestingTNewNetworkTrafficServiceClient interface {
+// NewNetworkTrafficServiceClient creates a new instance of NetworkTrafficServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewNetworkTrafficServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewNetworkTrafficServiceClient creates a new instance of NetworkTrafficServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewNetworkTrafficServiceClient(t mockConstructorTestingTNewNetworkTrafficServiceClient) *NetworkTrafficServiceClient {
+}) *NetworkTrafficServiceClient {
 	mock := &NetworkTrafficServiceClient{}
 	mock.Mock.Test(t)
 
