@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type NetworkTrafficServiceServer struct {
 // Modified provides a mock function with given fields: _a0, _a1
 func (_m *NetworkTrafficServiceServer) Modified(_a0 context.Context, _a1 *v1.NetworkTrafficReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *NetworkTrafficServiceServer) Modified(_a0 context.Context, _a1 *v1.Net
 func (_m *NetworkTrafficServiceServer) Usage(_a0 context.Context, _a1 *v1.NetworkUsageRequest) (*v1.NetworkUsageResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.NetworkUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.NetworkUsageRequest) (*v1.NetworkUsageResponse, error)); ok {
@@ -66,13 +74,12 @@ func (_m *NetworkTrafficServiceServer) Usage(_a0 context.Context, _a1 *v1.Networ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewNetworkTrafficServiceServer interface {
+// NewNetworkTrafficServiceServer creates a new instance of NetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewNetworkTrafficServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewNetworkTrafficServiceServer creates a new instance of NetworkTrafficServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewNetworkTrafficServiceServer(t mockConstructorTestingTNewNetworkTrafficServiceServer) *NetworkTrafficServiceServer {
+}) *NetworkTrafficServiceServer {
 	mock := &NetworkTrafficServiceServer{}
 	mock.Mock.Test(t)
 
