@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -27,6 +27,10 @@ func (_m *PodServiceClient) Added(ctx context.Context, in *v1.PodReport, opts ..
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Added")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -61,6 +65,10 @@ func (_m *PodServiceClient) Deleted(ctx context.Context, in *v1.PodReport, opts 
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Deleted")
+	}
+
 	var r0 *v1.Empty
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.PodReport, ...grpc.CallOption) (*v1.Empty, error)); ok {
@@ -93,6 +101,10 @@ func (_m *PodServiceClient) Modified(ctx context.Context, in *v1.PodReport, opts
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -127,6 +139,10 @@ func (_m *PodServiceClient) Usage(ctx context.Context, in *v1.ContainerUsageRequ
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.ContainerUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ContainerUsageRequest, ...grpc.CallOption) (*v1.ContainerUsageResponse, error)); ok {
@@ -149,13 +165,12 @@ func (_m *PodServiceClient) Usage(ctx context.Context, in *v1.ContainerUsageRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPodServiceClient interface {
+// NewPodServiceClient creates a new instance of PodServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPodServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPodServiceClient creates a new instance of PodServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPodServiceClient(t mockConstructorTestingTNewPodServiceClient) *PodServiceClient {
+}) *PodServiceClient {
 	mock := &PodServiceClient{}
 	mock.Mock.Test(t)
 
