@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type PodServiceServer struct {
 // Added provides a mock function with given fields: _a0, _a1
 func (_m *PodServiceServer) Added(_a0 context.Context, _a1 *v1.PodReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Added")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *PodServiceServer) Added(_a0 context.Context, _a1 *v1.PodReport) (*v1.E
 func (_m *PodServiceServer) Deleted(_a0 context.Context, _a1 *v1.PodReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Deleted")
+	}
+
 	var r0 *v1.Empty
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.PodReport) (*v1.Empty, error)); ok {
@@ -69,6 +77,10 @@ func (_m *PodServiceServer) Deleted(_a0 context.Context, _a1 *v1.PodReport) (*v1
 // Modified provides a mock function with given fields: _a0, _a1
 func (_m *PodServiceServer) Modified(_a0 context.Context, _a1 *v1.PodReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -96,6 +108,10 @@ func (_m *PodServiceServer) Modified(_a0 context.Context, _a1 *v1.PodReport) (*v
 func (_m *PodServiceServer) Usage(_a0 context.Context, _a1 *v1.ContainerUsageRequest) (*v1.ContainerUsageResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.ContainerUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.ContainerUsageRequest) (*v1.ContainerUsageResponse, error)); ok {
@@ -118,13 +134,12 @@ func (_m *PodServiceServer) Usage(_a0 context.Context, _a1 *v1.ContainerUsageReq
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPodServiceServer interface {
+// NewPodServiceServer creates a new instance of PodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPodServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPodServiceServer creates a new instance of PodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPodServiceServer(t mockConstructorTestingTNewPodServiceServer) *PodServiceServer {
+}) *PodServiceServer {
 	mock := &PodServiceServer{}
 	mock.Mock.Test(t)
 
