@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type PostgresServiceServer struct {
 // Added provides a mock function with given fields: _a0, _a1
 func (_m *PostgresServiceServer) Added(_a0 context.Context, _a1 *v1.PostgresReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Added")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *PostgresServiceServer) Added(_a0 context.Context, _a1 *v1.PostgresRepo
 func (_m *PostgresServiceServer) Deleted(_a0 context.Context, _a1 *v1.PostgresReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Deleted")
+	}
+
 	var r0 *v1.Empty
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.PostgresReport) (*v1.Empty, error)); ok {
@@ -69,6 +77,10 @@ func (_m *PostgresServiceServer) Deleted(_a0 context.Context, _a1 *v1.PostgresRe
 // Modified provides a mock function with given fields: _a0, _a1
 func (_m *PostgresServiceServer) Modified(_a0 context.Context, _a1 *v1.PostgresReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -96,6 +108,10 @@ func (_m *PostgresServiceServer) Modified(_a0 context.Context, _a1 *v1.PostgresR
 func (_m *PostgresServiceServer) Usage(_a0 context.Context, _a1 *v1.PostgresUsageRequest) (*v1.PostgresUsageResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.PostgresUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.PostgresUsageRequest) (*v1.PostgresUsageResponse, error)); ok {
@@ -118,13 +134,12 @@ func (_m *PostgresServiceServer) Usage(_a0 context.Context, _a1 *v1.PostgresUsag
 	return r0, r1
 }
 
-type mockConstructorTestingTNewPostgresServiceServer interface {
+// NewPostgresServiceServer creates a new instance of PostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewPostgresServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewPostgresServiceServer creates a new instance of PostgresServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewPostgresServiceServer(t mockConstructorTestingTNewPostgresServiceServer) *PostgresServiceServer {
+}) *PostgresServiceServer {
 	mock := &PostgresServiceServer{}
 	mock.Mock.Test(t)
 
