@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -27,6 +27,10 @@ func (_m *S3ServiceClient) BucketStats(ctx context.Context, in *v1.S3BucketRepor
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for BucketStats")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -61,6 +65,10 @@ func (_m *S3ServiceClient) Usage(ctx context.Context, in *v1.S3UsageRequest, opt
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.S3UsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.S3UsageRequest, ...grpc.CallOption) (*v1.S3UsageResponse, error)); ok {
@@ -83,13 +91,12 @@ func (_m *S3ServiceClient) Usage(ctx context.Context, in *v1.S3UsageRequest, opt
 	return r0, r1
 }
 
-type mockConstructorTestingTNewS3ServiceClient interface {
+// NewS3ServiceClient creates a new instance of S3ServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewS3ServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewS3ServiceClient creates a new instance of S3ServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewS3ServiceClient(t mockConstructorTestingTNewS3ServiceClient) *S3ServiceClient {
+}) *S3ServiceClient {
 	mock := &S3ServiceClient{}
 	mock.Mock.Test(t)
 
