@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type S3ServiceServer struct {
 // BucketStats provides a mock function with given fields: _a0, _a1
 func (_m *S3ServiceServer) BucketStats(_a0 context.Context, _a1 *v1.S3BucketReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for BucketStats")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *S3ServiceServer) BucketStats(_a0 context.Context, _a1 *v1.S3BucketRepo
 func (_m *S3ServiceServer) Usage(_a0 context.Context, _a1 *v1.S3UsageRequest) (*v1.S3UsageResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.S3UsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.S3UsageRequest) (*v1.S3UsageResponse, error)); ok {
@@ -66,13 +74,12 @@ func (_m *S3ServiceServer) Usage(_a0 context.Context, _a1 *v1.S3UsageRequest) (*
 	return r0, r1
 }
 
-type mockConstructorTestingTNewS3ServiceServer interface {
+// NewS3ServiceServer creates a new instance of S3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewS3ServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewS3ServiceServer creates a new instance of S3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewS3ServiceServer(t mockConstructorTestingTNewS3ServiceServer) *S3ServiceServer {
+}) *S3ServiceServer {
 	mock := &S3ServiceServer{}
 	mock.Mock.Test(t)
 
