@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafePodServiceServer) mustEmbedUnimplementedPodServiceServer() {
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafePodServiceServer interface {
+// NewUnsafePodServiceServer creates a new instance of UnsafePodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafePodServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafePodServiceServer creates a new instance of UnsafePodServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafePodServiceServer(t mockConstructorTestingTNewUnsafePodServiceServer) *UnsafePodServiceServer {
+}) *UnsafePodServiceServer {
 	mock := &UnsafePodServiceServer{}
 	mock.Mock.Test(t)
 
