@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -14,13 +14,12 @@ func (_m *UnsafeS3ServiceServer) mustEmbedUnimplementedS3ServiceServer() {
 	_m.Called()
 }
 
-type mockConstructorTestingTNewUnsafeS3ServiceServer interface {
+// NewUnsafeS3ServiceServer creates a new instance of UnsafeS3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewUnsafeS3ServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewUnsafeS3ServiceServer creates a new instance of UnsafeS3ServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewUnsafeS3ServiceServer(t mockConstructorTestingTNewUnsafeS3ServiceServer) *UnsafeS3ServiceServer {
+}) *UnsafeS3ServiceServer {
 	mock := &UnsafeS3ServiceServer{}
 	mock.Mock.Test(t)
 
