@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -27,6 +27,10 @@ func (_m *VolumeServiceClient) Added(ctx context.Context, in *v1.VolumeReport, o
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Added")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -61,6 +65,10 @@ func (_m *VolumeServiceClient) Deleted(ctx context.Context, in *v1.VolumeReport,
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Deleted")
+	}
+
 	var r0 *v1.Empty
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.VolumeReport, ...grpc.CallOption) (*v1.Empty, error)); ok {
@@ -93,6 +101,10 @@ func (_m *VolumeServiceClient) Modified(ctx context.Context, in *v1.VolumeReport
 	_ca = append(_ca, ctx, in)
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -127,6 +139,10 @@ func (_m *VolumeServiceClient) Usage(ctx context.Context, in *v1.VolumeUsageRequ
 	_ca = append(_ca, _va...)
 	ret := _m.Called(_ca...)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.VolumeUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.VolumeUsageRequest, ...grpc.CallOption) (*v1.VolumeUsageResponse, error)); ok {
@@ -149,13 +165,12 @@ func (_m *VolumeServiceClient) Usage(ctx context.Context, in *v1.VolumeUsageRequ
 	return r0, r1
 }
 
-type mockConstructorTestingTNewVolumeServiceClient interface {
+// NewVolumeServiceClient creates a new instance of VolumeServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewVolumeServiceClient(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewVolumeServiceClient creates a new instance of VolumeServiceClient. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewVolumeServiceClient(t mockConstructorTestingTNewVolumeServiceClient) *VolumeServiceClient {
+}) *VolumeServiceClient {
 	mock := &VolumeServiceClient{}
 	mock.Mock.Test(t)
 
