@@ -1,4 +1,4 @@
-// Code generated by mockery v2.22.1. DO NOT EDIT.
+// Code generated by mockery v2.38.0. DO NOT EDIT.
 
 package mocks
 
@@ -17,6 +17,10 @@ type VolumeServiceServer struct {
 // Added provides a mock function with given fields: _a0, _a1
 func (_m *VolumeServiceServer) Added(_a0 context.Context, _a1 *v1.VolumeReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Added")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -44,6 +48,10 @@ func (_m *VolumeServiceServer) Added(_a0 context.Context, _a1 *v1.VolumeReport) 
 func (_m *VolumeServiceServer) Deleted(_a0 context.Context, _a1 *v1.VolumeReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Deleted")
+	}
+
 	var r0 *v1.Empty
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.VolumeReport) (*v1.Empty, error)); ok {
@@ -69,6 +77,10 @@ func (_m *VolumeServiceServer) Deleted(_a0 context.Context, _a1 *v1.VolumeReport
 // Modified provides a mock function with given fields: _a0, _a1
 func (_m *VolumeServiceServer) Modified(_a0 context.Context, _a1 *v1.VolumeReport) (*v1.Empty, error) {
 	ret := _m.Called(_a0, _a1)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Modified")
+	}
 
 	var r0 *v1.Empty
 	var r1 error
@@ -96,6 +108,10 @@ func (_m *VolumeServiceServer) Modified(_a0 context.Context, _a1 *v1.VolumeRepor
 func (_m *VolumeServiceServer) Usage(_a0 context.Context, _a1 *v1.VolumeUsageRequest) (*v1.VolumeUsageResponse, error) {
 	ret := _m.Called(_a0, _a1)
 
+	if len(ret) == 0 {
+		panic("no return value specified for Usage")
+	}
+
 	var r0 *v1.VolumeUsageResponse
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context, *v1.VolumeUsageRequest) (*v1.VolumeUsageResponse, error)); ok {
@@ -118,13 +134,12 @@ func (_m *VolumeServiceServer) Usage(_a0 context.Context, _a1 *v1.VolumeUsageReq
 	return r0, r1
 }
 
-type mockConstructorTestingTNewVolumeServiceServer interface {
+// NewVolumeServiceServer creates a new instance of VolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewVolumeServiceServer(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewVolumeServiceServer creates a new instance of VolumeServiceServer. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewVolumeServiceServer(t mockConstructorTestingTNewVolumeServiceServer) *VolumeServiceServer {
+}) *VolumeServiceServer {
 	mock := &VolumeServiceServer{}
 	mock.Mock.Test(t)
 
